Fix migration files

+++ db/migrate/20250108222910_remove_mfacp_from_fire_alarm_control_panels.rb
@@ -1,5 +1,7 @@
 class RemoveMfacpFromFireAlarmControlPanels < ActiveRecord::Migration[7.1]
   def change
-    remove_column :fire_alarm_control_panels, :mfacp, :string
+    if column_exists?(:fire_alarm_control_panels, :mfacp)
+      remove_column :fire_alarm_control_panels, :mfacp, :string
+    end
   end
 end
