+++ db/migrate/20250113204552_remove_project_id_from_systems.rb
@@ -1,5 +1,7 @@
 class RemoveProjectIdFromSystems < ActiveRecord::Migration[7.1]
   def change
-    remove_column :systems, :project_id, :bigint
+    if column_exists?(:systems, :project_id)
+      remove_column :systems, :project_id, :bigint
+    end
   end
 end
