+++ db/migrate/20250127135536_drop_evacuation_systems.rb
@@ -1,5 +1,7 @@
 class DropEvacuationSystems < ActiveRecord::Migration[7.1]
   def change
-    drop_table :evacuation_systems
+    if table_exists?(:evacuation_systems)
+      drop_table :evacuation_systems
+    end
   end
 end
