+++ db/migrate/20250201141356_remove_subsystems_from_suppliers.rb
@@ -1,5 +1,7 @@
 class RemoveSubsystemsFromSuppliers < ActiveRecord::Migration[7.1]
   def change
-    remove_column :suppliers, :subsystems, :text
+    if column_exists?(:suppliers, :subsystems)
+      remove_column :suppliers, :subsystems, :text
+    end
   end
 end
