+++ db/migrate/20250201215202_remove_membership_type_from_suppliers.rb
@@ -1,5 +1,7 @@
 class RemoveMembershipTypeFromSuppliers < ActiveRecord::Migration[7.1]
   def change
-    remove_column :suppliers, :membership_type, :string
+    if column_exists?(:suppliers, :membership_type)
+      remove_column :suppliers, :membership_type, :string
+    end
   end
 end
